Correct a major bug in timer.c on 32 bits platform (at least).
Correct a major bug in timer.c on all the platforms where an
"unsigned long" cannot totally contain the current epoch (in ms).
An epoch in ms requires 64 bits and on the 32 bits platforms
"unsigned long" are 32 bits wide.
So, big time skips can cause integer overflows on the various timers
and create unattainable timeouts.

The test case was pretty simple:
1. Start a system in 2010.
2. Run an Ivy program which uses timers.
3. Update the clock to 2016.
4. Integer overflow occurs on the "current" time.
4. Ivy is blocked because the previous timeout is now very far.

This patch is not perfect but is a good starting point.